Allow delete attachments from builder data section

+++ packages/bbui/src/Form/Core/Dropzone.svelte
@@ -95,9 +95,11 @@
       "change",
       value.filter((x, idx) => idx !== selectedImageIdx)
     )
-    await deleteAttachments(
-      value.filter((x, idx) => idx === selectedImageIdx).map(item => item.key)
-    )
+    if (deleteAttachments) {
+      await deleteAttachments(
+        value.filter((x, idx) => idx === selectedImageIdx).map(item => item.key)
+      )
+    }
     selectedImageIdx = 0
   }
